+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
- 
 
- 
+## [3.3.0] - 2025-10-26
+
+### Added
+- System-wide installation feature (`install -is` or `install --install-system`) to install `huntools` to `/usr/local/bin`.
+- Documentation for the new system-wide installation feature in `README.md`.
+
+### Changed
+- The `remove -ca` (`clean_all`) command now also removes the system-wide `huntools` executable if it exists.
+
 ## [3.2.0] - 2025-10-26
 
 ### Added